Atualização dos campos de listar instrumentos e quantidade, na view.
