+++ Codigo/Core/DTO/EventoDTO.cs
@@ -22,29 +22,4 @@ namespace Core.DTO
         [Display(Name = "Confirmados")]
         public int Confirmados { get; set; }
     }
-
-
-    public class GerenciarInstrumentoEventoDTO
-    {
-        public int Id { get; set; }
-
-        [Display(Name = "Data hora início")]
-        public DateTime DataHoraInicio { get; set; }
-
-        [Display(Name = "Data hora fim")]
-        public DateTime DataHoraFim { get; set; }
-
-        [Display(Name = "Regente")]
-        public IEnumerable<string>? Regente { get; set; }
-
-        [Display(Name = "Figurino")]
-        public IEnumerable<string>? Figurino { get; set; }
-
-        [Display(Name = "Local")]
-        public string? Local { get; set; }
-
-        [Display(Name = "Instrumento")]
-        public IEnumerable<Tipoinstrumento>? Instrumento { get; set; }
-
-    }
 }
