+++ Codigo/Core/DTO/MovimentacaoInstrumentoDTO.cs
@@ -1,5 +1,4 @@
-﻿
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
 namespace Core.DTO
 {
