+++ Codigo/GestaoGrupoMusicalWeb/Views/Evento/GerenciarInstrumentoEvento.cshtml
@@ -2,7 +2,7 @@
 @{
     ViewData["Title"] = "Gerenciar Instrumento Evento";
     ViewData["Evento"] = "Eventos";
-    ViewData["GerenciarInstrumentoEventoViewModel"] = "Gerenciar Instrumento";
+    ViewData["GerenciarInstrumentoEvento"] = "Gerenciar Instrumento";
 }
 
 <partial name="_Notificar">
@@ -89,15 +89,13 @@
 
                 <div class="row">
                     <div class="d-flex justify-content-lg-end flex-column-reverse flex-lg-row">
-                        <a class="btn btn-outline-secondary col-lg-2 mt-3 me-lg-3 btn-" asp-controller="Figurino" asp-action="Index">Voltar</a>                        
+                        <a class="btn btn-outline-secondary col-lg-2 mt-3 me-lg-3 btn-" asp-controller="Evento" asp-action="Index">Voltar</a>                        
                     </div>
                 </div>
         </div>
        
                         
     </div>
-
-
 
     @section Header {
         @{
